Moved QApplication into one test
